Working Arm Code
It has the correct ports for the TalonSRX and SparkMax. The comments are correct for moving the arm up and down. There is commented out code that will be used to implement PID. I just need to add commands as soon as we are sure everything is working(the intake).